fix(geo): validate positive count arguments

+++ src/types/redis_geo.h
@@ -68,16 +68,16 @@ class Geo : public ZSet {
   rocksdb::Status Pos(engine::Context &ctx, const Slice &user_key, const std::vector<Slice> &members,
                       std::map<std::string, GeoPoint> *geo_points);
   rocksdb::Status Radius(engine::Context &ctx, const Slice &user_key, double longitude, double latitude,
-                         double radius_meters, int count, DistanceSort sort, const std::string &store_key,
+                         double radius_meters, size_t count, DistanceSort sort, const std::string &store_key,
                          bool store_distance, double unit_conversion, std::vector<GeoPoint> *geo_points);
   rocksdb::Status RadiusByMember(engine::Context &ctx, const Slice &user_key, const Slice &member, double radius_meters,
-                                 int count, DistanceSort sort, const std::string &store_key, bool store_distance,
+                                 size_t count, DistanceSort sort, const std::string &store_key, bool store_distance,
                                  double unit_conversion, std::vector<GeoPoint> *geo_points);
   rocksdb::Status Search(engine::Context &ctx, const Slice &user_key, GeoShape geo_shape, OriginPointType point_type,
-                         std::string &member, int count, DistanceSort sort, bool store_distance, double unit_conversion,
-                         std::vector<GeoPoint> *geo_points);
+                         std::string &member, size_t count, DistanceSort sort, bool store_distance,
+                         double unit_conversion, std::vector<GeoPoint> *geo_points);
   rocksdb::Status SearchStore(engine::Context &ctx, const Slice &user_key, GeoShape geo_shape,
-                              OriginPointType point_type, std::string &member, int count, DistanceSort sort,
+                              OriginPointType point_type, std::string &member, size_t count, DistanceSort sort,
                               const std::string &store_key, bool store_distance, double unit_conversion,
                               std::vector<GeoPoint> *geo_points);
   rocksdb::Status Get(engine::Context &ctx, const Slice &user_key, const Slice &member, GeoPoint *geo_point);
